requirements.txt
openai>=1.3.0
openai-whisper>=20231117
pyttsx3>=2.90
requests>=2.31.0
sounddevice>=0.4.6
soundfile>=0.12.1
torch>=2.0.0
numpy>=1.24.0
